requirements.txt
numpy==1.16.2
scipy==1.2.1
tensorflow-gpu==1.15.2
argparse
parse
tqdm
open3d-python
